完成域间权限的增加和获取

+++ fabricPlatform/api-commons/src/main/java/com/hust/keyRD/commons/entities/ChannelDataAuthority.java
@@ -24,22 +24,22 @@ public class ChannelDataAuthority implements Serializable {
     /**
      * channel_data_authority_id
      */
-    private Long id;
+    private Integer id;
 
     /**
      * 发送者或拉取者id
      */
-    private Long userId;
+    private Integer userId;
 
     /**
      * 文件id
      */
-    private Long dataId;
+    private Integer dataId;
 
     /**
      * channelID
      */
-    private Long channelId;
+    private Integer channelId;
 
     /**
      * 权限类型：1-push，2-pull
